Refactors. Improves thread safety. Adds type check. Removes temporary code.
