+++ src/stream-chat-net/JSONHelpers.cs
@@ -1,119 +1,56 @@
-using System;
-using System.Linq;
-using System.Collections.Generic;
-using System.Reflection;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
-using System.Collections;
+using System.Reflection;
 
 namespace StreamChat
 {
-    public class CustomDataBase
-    {
-        protected GenericData _data = new GenericData();
-
-        public CustomDataBase() { }
-
-        public T GetData<T>(string name)
-        {
-            return this._data.GetData<T>(name);
-        }
-
-        public void SetData<T>(string name, T data)
-        {
-            this._data.SetData<T>(name, data);
-        }
-
-        internal JObject ToJObject()
-        {
-            var root = JObject.FromObject(this);
-            this._data.AddToJObject(root);
-            return root;
-        }
-    }
-
-    internal class JsonHelpers
+    internal static class JsonHelpers
     {
         private delegate GenericData Builder<T>(T target, IEnumerable<JProperty> jsonProps);
 
-        private static readonly Dictionary<Type, object> CachedBuilders = new Dictionary<Type, object>();
+        private static readonly ConcurrentDictionary<Type, object> CachedBuilders = new ConcurrentDictionary<Type, object>();
 
         internal static GenericData FromJObject<T>(T obj, JObject json)
+            where T : class
         {
             var jProps = json.Properties();
-
-            if (CachedBuilders.TryGetValue(typeof(T), out var builder))
-                return (builder as Builder<T>)(obj, jProps);
-
-            var properties = typeof(T).GetPropertiesPortable();
-
-            var newBuilder = ConstructBuilder<T>(properties);
-
-            CachedBuilders[typeof(T)] = newBuilder;
-
-            return newBuilder(obj, jProps);
+            
+            var builder = (Builder<T>)CachedBuilders.GetOrAdd(typeof(T), _ => ConstructBuilder<T>());
+            
+            return builder(obj, jProps);
         }
 
         internal static void RegisterType<T>()
+            where T : class => 
+            CachedBuilders.GetOrAdd(typeof(T), _ => ConstructBuilder<T>());
+
+        private static Builder<T> ConstructBuilder<T>()
         {
-            if (CachedBuilders.ContainsKey(typeof(T)))
-                return;
+            var properties = typeof(T)
+                .GetPropertiesPortable()
+                .Where(p => p.SetMethod != null);
 
-            var properties = typeof(T).GetPropertiesPortable();
-
-            var newBuilder = ConstructBuilder<T>(properties);
-
-            CachedBuilders[typeof(T)] = newBuilder;
-        }
-
-        private static Builder<T> ConstructBuilder<T>(PropertyInfo[] properties)
-        {
             var mappedProperties = GetObjectPropertyMap(properties);
 
             var extraDataExp = Expression.Variable(typeof(GenericData), "extra");
             var targetExp = Expression.Parameter(typeof(T), "target");
             var jsonPropertiesExp = Expression.Parameter(typeof(IEnumerable<JProperty>), "jProps");
-            var jsonPropertyType = typeof(JProperty);
 
             var forEachExp = ReflectionHelper.CreateForEach(
                     jsonPropertiesExp,
-                    currentExp =>
-                    {
-                        var currentValueExp = Expression.Property(currentExp, jsonPropertyType.GetPropertyPortable(nameof(JProperty.Value)));
-                        var currentNameExp = Expression.Property(currentExp, jsonPropertyType.GetPropertyPortable(nameof(JProperty.Name)));
+                    current => CreateForEachBody(current, mappedProperties, extraDataExp, targetExp));
 
-                        var defaultCase = Expression.Call(
-                                extraDataExp,
-                                typeof(GenericData).GetMethodPortable(nameof(GenericData.SetData)).MakeGenericMethod(typeof(JToken)),
-                                currentNameExp,
-                                currentValueExp);
-
-                        var cases = mappedProperties
-                                .Select(pair =>
-                                {
-                                    var currentValueToObject = Expression.Call(
-                                       currentValueExp,
-                                       jsonPropertyType
-                                           .GetMethodPortable(nameof(JProperty.ToObject), Type.EmptyTypes)
-                                           .MakeGenericMethod(pair.Value.PropertyType));
-
-                                    return Expression.SwitchCase(
-                                        Expression.Assign(Expression.Property(targetExp, pair.Value), currentValueToObject),
-                                        Expression.Constant(pair.Key));
-
-                                })
-                                .ToArray();
-
-                        var result = Expression.Switch(typeof(void), currentNameExp, defaultCase, null, cases);
-
-                        return result;
-                    });
+            var initExtraDataExp = Expression.Assign(extraDataExp, Expression.New(typeof(GenericData)));
 
             var finalExpression = Expression.Lambda<Builder<T>>(
                 Expression.Block(
                     new[] { extraDataExp },
-                    Expression.Assign(extraDataExp, Expression.New(typeof(GenericData))),
+                    initExtraDataExp,
                     forEachExp,
                     extraDataExp),
                 targetExp,
@@ -123,7 +60,44 @@ namespace StreamChat
             return finalExpression.Compile();
         }
 
-        private static Dictionary<string, PropertyInfo> GetObjectPropertyMap(PropertyInfo[] properties)
+        private static Expression CreateForEachBody(
+            Expression currentExp, 
+            Dictionary<string, PropertyInfo> mappedProperties, 
+            Expression extraDataExp, 
+            Expression targetExp)
+        {
+            var jsonPropertyType = typeof(JProperty);
+            var currentValueExp = Expression.Property(currentExp, jsonPropertyType.GetPropertyPortable(nameof(JProperty.Value)));
+            var currentNameExp = Expression.Property(currentExp, jsonPropertyType.GetPropertyPortable(nameof(JProperty.Name)));
+
+            var defaultCase = Expression.Call(
+                    extraDataExp,
+                    typeof(GenericData).GetMethodPortable(nameof(GenericData.SetData)).MakeGenericMethod(typeof(JToken)),
+                    currentNameExp,
+                    currentValueExp);
+
+            var cases = mappedProperties
+                    .Select(pair =>
+                    {
+                        var currentValueToObject = Expression.Call(
+                            currentValueExp,
+                            jsonPropertyType
+                                .GetMethodPortable(nameof(JProperty.ToObject), Type.EmptyTypes)
+                                .MakeGenericMethod(pair.Value.PropertyType));
+
+                        return Expression.SwitchCase(
+                            Expression.Assign(Expression.Property(targetExp, pair.Value), currentValueToObject),
+                            Expression.Constant(pair.Key));
+
+                    })
+                    .ToArray();
+
+            var result = Expression.Switch(typeof(void), currentNameExp, defaultCase, null, cases);
+
+            return result;
+        }
+
+        private static Dictionary<string, PropertyInfo> GetObjectPropertyMap(IEnumerable<PropertyInfo> properties)
         {
             var mappedProperties = new Dictionary<string, PropertyInfo>();
 
@@ -140,7 +114,8 @@ namespace StreamChat
                         propertyName = propertyAttribute.PropertyName;
                         break;
                     }
-                    else if (attribute is JsonIgnoreAttribute ignoreAttr)
+                    
+                    if (attribute is JsonIgnoreAttribute ignoreAttr)
                     {
                         ignore = true;
                         break;
@@ -154,144 +129,6 @@ namespace StreamChat
             }
 
             return mappedProperties;
-        }
-
-        internal static GenericData FromJObjectOld<T>(T obj, JObject json)
-        {
-            var properties = typeof(T).GetPropertiesPortable();
-
-            var objProps = new Dictionary<string, PropertyInfo>();
-            var extra = new GenericData();
-
-            foreach (var prop in properties)
-            {
-                bool ignore = false;
-                string propName = prop.Name;
-                var attrs = prop.GetCustomAttributes(true);
-                foreach (var attr in attrs)
-                {
-                    JsonIgnoreAttribute ignoreAttr = attr as JsonIgnoreAttribute;
-                    if (ignoreAttr != null)
-                    {
-                        ignore = true;
-                        break;
-                    }
-                    JsonPropertyAttribute result = attr as JsonPropertyAttribute;
-                    if (result != null)
-                    {
-                        propName = result.PropertyName;
-                        break;
-                    }
-                }
-                if (!ignore && !objProps.ContainsKey(propName))
-                    objProps.Add(propName, prop);
-            }
-            var jsonProps = json.Properties();
-            foreach (var jsonProp in jsonProps)
-            {
-                PropertyInfo objProp;
-                if (objProps.TryGetValue(jsonProp.Name, out objProp))
-                    objProp.SetValue(obj, jsonProp.Value.ToObject(objProp.PropertyType));
-                else
-                    extra.SetData(jsonProp.Name, jsonProp.Value);
-            }
-            return extra;
-        }
-    }
-
-    internal static class ReflectionHelper
-    {
-        public static PropertyInfo GetPropertyPortable(this Type type, string propName)
-        {
-            return type
-#if NETSTANDARD1_6
-                .GetTypeInfo()
-#endif
-                .GetProperty(propName);
-        }
-
-        public static PropertyInfo[] GetPropertiesPortable(this Type type)
-        {
-            return type
-#if NETSTANDARD1_6
-                .GetTypeInfo()
-#endif
-                .GetProperties();
-        }
-
-        public static MethodInfo GetMethodPortable(this Type type, string methodName)
-        {
-            return type
-#if NETSTANDARD1_6
-                .GetTypeInfo()
-#endif
-                .GetMethod(methodName);
-        }
-
-        public static MethodInfo GetMethodPortable(this Type type, string methodName, params Type[] types)
-        {
-            types = types ?? (new Type[0]);
-
-            return type
-#if NETSTANDARD1_6
-                .GetTypeInfo()
-#endif
-                .GetMethod(methodName, types);
-        }
-
-        public static Type[] GetInterfacesPortable(this Type type)
-        {
-            return type
-#if NETSTANDARD1_6
-                .GetTypeInfo()
-#endif
-                .GetInterfaces();
-        }
-
-        public static Expression CreateForEach(ParameterExpression enumerableExp, Func<ParameterExpression, Expression> bodyGenerator)
-        {
-            var breakLabel = Expression.Label("ForeachBreak");
-            var getEnumeratorMethod = enumerableExp.Type.GetMethodPortable(nameof(IEnumerable.GetEnumerator));
-            var enumeratorType = getEnumeratorMethod.ReturnType;
-            var enumeratorExp = Expression.Parameter(enumeratorType, "enumerator");
-            var moveNextMethod = typeof(IEnumerator).GetMethodPortable(nameof(IEnumerator.MoveNext));
-            var currentProperty = enumeratorType.GetPropertyPortable(nameof(IEnumerator.Current));
-            var currentExpression = Expression.Parameter(currentProperty.PropertyType, "current");
-
-            Expression loopBlock = Expression.Loop(
-                        Expression.IfThenElse(
-                            Expression.Call(enumeratorExp, moveNextMethod),
-                            Expression.Block(
-                                new[] { currentExpression },
-                                Expression.Assign(
-                                    currentExpression,
-                                    Expression.Property(enumeratorExp, currentProperty)),
-                                bodyGenerator(currentExpression)
-                                ),
-                            Expression.Break(breakLabel)));
-
-            var diposableType = typeof(IDisposable);
-
-            if (enumeratorType.GetInterfacesPortable().Contains(diposableType))
-            {
-                var disposeMethod = diposableType.GetMethodPortable(nameof(IDisposable.Dispose));
-
-                loopBlock = Expression.TryFinally(loopBlock, Expression.Call(Expression.Convert(enumerableExp, diposableType), disposeMethod));
-            }
-
-            var res = Expression.Block(
-                    new[]
-                    {
-                        enumeratorExp
-                    },
-                    Expression.Assign(
-                        enumeratorExp,
-                        Expression.Call(enumerableExp, getEnumeratorMethod)),
-                    loopBlock,
-                    Expression.Label(breakLabel)
-                );
-
-            return res;
         }
     }
 }
